requirements.txt
# Core LangGraph and LangChain
langgraph>=0.0.55
langchain>=0.1.0
langchain-openai>=0.1.0

# Web framework and server
fastapi>=0.104.0
uvicorn>=0.24.0

# Data validation and HTTP client
pydantic>=2.0.0
httpx>=0.25.0

# Environment and utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Memory integration dependencies
aiofiles>=23.2.0
asyncio-throttle>=1.0.2

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Additional utilities
python-json-logger>=2.0.7
tenacity>=8.2.0